Add dartsass build to precompile assets task
In theory the Rails dartsass gem should automatically link the dartsass build command to the precomile assets rake task. In practice that doesn't seem to work, perhaps because we're doing our own enhancement to add the yarn install command. This commit ensures that the dart sass build runs during the precompilation of assets