Allow unpublsihing when a manual state is withdrawn
We make sure the publishing-api call succeeds before changing the state of the manual.
We allow the unpublish call to execute when the manual is either published or (already) withdrawn.

We have experienced that we are unable to rerun the `withdraw_and_redirect_manual` rake task, should the
unpublish api call fail. These changes ensure we can rerun.

Co-authored-by: Minno Dang <minno.dang@digital.cabinet-office.gov.uk>