Remove rake task to withdraw and redirect manual and sections to different paths via a CSV file
This rake task is forcing us to keep the flag and behaviour where we withdraw and redirect the manual without its sections. since we want to redirect them to different places. This is causing risk where people could be misusing the flag and we end up with a manual that is withdrawn but its sections are not.
In cases like this, we suggest to unpublish and withdraw the manual and its sections and if there are any specific URL redirects, to use the short URL manager for this.