Add rake task for changing org slug for manual records
The "Change an organisation's slug" instructions, at time of
writing, ask that we create a 'migration' in Manuals Publisher.
https://docs.publishing.service.gov.uk/manual/changing-organisation-slug.html#2-update-the-organisation-slug-in-manuals-publisher

But as per previous commit, rails migrations haven't been supported
in this app since 2020, preferring 'temporary rake tasks'. This
commit therefore adds a rake task, though the intention is to
keep it, since we'll need to run it whenever an organisation slug
is changed.

The slug is only displayed in Manuals Publisher itself (and used
to determine which manuals the user can see), so no republishing
is required as part of this task.