Bulk withdraw and redirect orphaned sections
In order to address some data integrity issues where we have withdrawn Manuals but the children sections are published. We need to be able to go through a long list of sections and unpublish them from Publishing API. This rake task should be temporary and run one-off since we don't expect we'll have to do this again in the future.