Fixing rake task so it sends discard draft correctly to Publishing API
Previously, the raks task for removing sections had a conditional where it only sent discard draft only if the section was a draft and had the discard draft flag. However, we have sections that are currently published which have drafts on Publishing API - we are unable to remove them due to an error "Unable to unpublish edition with draft". Amending this so we can unpublish these sections.

Additionally, a few of the URLs were wrong, and required regenerating.