When withdrawing sections using the rake task, only look for published sections
There are some strange data in the database whilst investigating published section pages that have withdrawn manuals. Three of these sections failed to be unpublished because when looking up by slug, there are 3 editions to the section, the latest of which is a draft edition with a DIFFERENT section_uuid to its predecessors. Looking at other sections, all of the editions should be sharing a section uuid - this leads to a bug where we can't trace back to the previous published edition via the draft. To get around this, we should be looking for published editions only, since that's what we check for eventually anyways.