Fixing withdraw section rake task so it sends discard draft directly through to Publishing API
Previously, the rake task for removing sections had a conditional where it only sent discard draft flag to Publishing API only if the latest section edition was a draft and the rake task had the discard draft flag set to true.

However, when the rake task was run to remove orphaned sections, we seem to have sections where the latest edition is published (not draft) but somehow contain drafts on Publishing API - we were unable to remove them due to an error "Unable to unpublish edition with draft". Amending the publishing API call for this Withdraw Section task to always send discard_draft flag straight through rather than check for draft so we can unpublish these sections.