Withdraw sections by slug should only look up published editions
There are some strange data in the database. Whilst investigating withdrawn manuals that have published sections, three of these sections failed to be withdrawn. Reason is when looking up SectionEdition by slug, there were 3 editions that were returned, the latest of which is a draft edition. This draft edition has a DIFFERENT section_uuid to its predecessors. Looking at other sections, all of the editions should be sharing a section uuid - this leads to a bug where the Section.find method can't trace back to the previous published edition via the draft, meaning edition.published? is returning an incorrect value, and the withdraw task fails.

To get around this, we should be looking up published editions only, since that's what we check for eventually anyways.