Remove adapters file
All the adapters file is doing is instantiating OrganisationsAdapter and PublishingAdapter - both of which contain methods with static method behaviour, with no instance specific state. Instead of having to jump over two levels to get to the method, directly calling the respective adapters might be the quickest way.
The refactor also comes with some rewriting of certain tests, since there is a lot of low level mocking which the restructure breaks. The most significant change in tests is the Section Removal Service spec where it is updated to use real data models like manuals and sections rather than stub out all the calls to them. With the complex logic that goes on, it would make more sense to test what actually happens