Fix error handling for when Manual Not Found
Added a lot of missing test cases for when the manual is not found. From these test cases we can see that when a manual is not found, it throws Manual::NotFoundError, not a key error, so all the rescues for those will not be firing, making all ManualNotFoundErrors redundant. Cleaning all of those up.