Move the responsibility for generating slugs outside of the section
Section slugs depend on the manual, which lead to an awkward dependency where the manual had to be passed to the section for it to initialise itself.

This made code to load a section directly, which we do in a few places such as when redirecting a section, more awkward than necessary, because it was necessary to load a manual to instantiate a section even though the section already had a slug.

It therefore makes sense to move slug generation outside of the section and remove the need for sections to depend directly on manuals. I have chosen to move the slug generation process into the section service layer.

Ideally I think the manual would act as an aggregate for sections and we would not manipulate sections directly other than through the manual, but that would have been a much bigger refactor. If we took that route then the manual could be responsible for reslugging sections.