Move documentation from the 2017 database migration project into an archive
I have decided to move the existing project documentation into an archive. The existing documentation was written from the point of view of the team that considered a migration project in 2017 that was never completed.

Because the documentation was written as a record of the system at a point in time, rather than as an attempt to provide live service documentation, it is confusing when trying to understand the current application, and some of it is out of date. However, many parts of the documentation are still useful, particularly from the perspective of understanding why certain decisions have been made. I therefore believe it would be valuable to keep it in the repository.

I will write a new set of documentation for the service in the near future.