Fix rake task to simply redirect rather than changing any state on the section when redirecting section
Upon further investigation, there is no syncing between document state in Manuals Publisher and Publishing API. The sections are meant to be left at "published" state even when their manual parents have been withdrawn. The redirects that are breaking in production are introduced because they are redirects set on Router and not Publishing API unpublish/redirects.
This temporary rake task should not make any changes to the state, but just push the redirects.

Additionally added a check for safety when withdrawing sections to make sure we are fetching the correct UUID for a published section. If there are multiple published section editions with differing Section UUIDs, this should be a cause for concern and warrant further investigation. We error out instead of actioning the task.