Expect discard draft button to be visible after saving manual or section
Some features are occasionally failing because visiting the manual page fails after saving a manual or section.

According to https://github.com/rubygems/rubygems.org/pull/5542, this occurs because previous requests have not finished when the next one starts.

We therefore add an extra assert to check that we have completed navigation to the show manual page. Without having access to the original manual, the best I could think of to do was to look for the presence of the discard draft button.