Add confirmation step to reslug_organisation rake task
I have added a confirmation step, using the Thor gem, to the above rake task. This is to ensure that the person running these rake tasks explicitly opts in to making these changes. I have also added a stub in the before :each block in reslug_organisation_spec to prevent the rake task from waiting for actual user input during the test.